Disable a few tests
